requirements.txt
tensorflow==2.17.0
transformers
sentence-transformers
ultralytics
opencv-python
scikit-learn
matplotlib
pandas
thop
python-multipart
streamlit
ipykernel
fastapi
spacy
huggingface_hub
uvicorn
faiss-cpu
tf-keras
python-dotenv
openai
groq
